+++ pyproject.toml
@@ -21,6 +21,7 @@ dependencies = [
   "beautifulsoup4==4.14.2",
   "brotli==1.1.0",
   "calmjs.parse==1.3.3",
+  "cryptography==46.0.7",
   "csscompressor==0.9.5",
   "deepl==1.23.0",
   "diff-match-patch==20241021",
@@ -58,6 +59,7 @@ dependencies = [
   "hiredis==3.3.0",
   "holidays==0.84",
   "html5lib==1.1",
+  "httpx==0.28.1",
   "humanize==4.14.0",
   "inflection==0.5.1",
   "jellyfish==1.2.1",
@@ -91,7 +93,6 @@ dependencies = [
   "requests==2.32.5",
   "requests-oauthlib==2.0.0",
   "ruff==0.14.3",
-  "satispaython==0.4.0",
   "setuptools==80.9.0",
   "slugify==0.0.1",
   "stripe==13.1.1",
